coding at max game

+++ DonutStoreUpgrade/src/main/java/ds/upgrade/util/UIUtil.java
@@ -3,7 +3,8 @@ package ds.upgrade.util;
 import java.math.BigInteger;
 
 public class UIUtil {
-  public static String formatNumber(BigInteger number) {
-    return String.format("%,d", number);
+  public static String formatNumber(Long number) {
+	  
+    return String.format("%,d",BigInteger.valueOf(number));
   }
 }
